SUIT hands-on notebooks for ESA-ISRO workshop.
The coalignment_module gives a detailed tutorial on how to coalign different SUIT filters. lc_template gives a tutorial on how to create lightcurves using SUIT data. This template relies on an understanding of the first module. environment.txt is just the environment file, and it will work for both notebooks.